+++ examples/http/client.js
@@ -16,19 +16,31 @@ function obtainCharset (headers) {
 }
 
 function pyro_call(object, method, callback) {
-    http.get("http://localhost:8080/pyro/"+object+"/"+method, function(res) {
-        var charset = obtainCharset(res.headers);
-        res.setEncoding(charset);
-        buffer='';
-        res.on('data', function(d) {
-            buffer += d.toString();
+    http.get({
+            hostname: "localhost",
+            port: 8080,
+            path: "/pyro/"+object+"/"+method,
+            headers: {
+                    // "X-Pyro-Options": "oneway"
+                }
+            },
+        function(res) {
+            var charset = obtainCharset(res.headers);
+            res.setEncoding(charset);
+            buffer='';
+            res.on('data', function(d) {
+                buffer += d.toString();
         });
 
         res.on('end', function() {
             if(res.statusCode==200) {
                 // all was well, process the response data as json
-                var parsed = JSON.parse(buffer);
-                callback(parsed);
+                if(buffer) {
+                    var parsed = JSON.parse(buffer);
+                    callback(parsed);
+                } else {
+                    callback(null);
+                }
             } else {
                 // 404, 500 or some other server error occurred
                 console.error("Server returned error response:");
@@ -54,5 +66,15 @@ pyro_call("Pyro.NameServer", "$meta", function(response) {
 
 pyro_call("Pyro.NameServer", "lookup?name=Pyro.NameServer", function(response) { 
     console.log("\nLOOKUP--->");
+    console.log(JSON.stringify(response, null, 4)); 
+});
+
+pyro_call("test.echoserver", "oneway_slow", function(response) { 
+    console.log("\nONEWAY_SLOW--->");
+    console.log(JSON.stringify(response, null, 4)); 
+});
+
+pyro_call("test.echoserver", "slow", function(response) { 
+    console.log("\nSLOW--->");
     console.log(JSON.stringify(response, null, 4)); 
 });